refactor: 826 - move route.types.ts types to shared types.ts
