--- src/core/types/route.types.ts
@@ -1,5 +1,0 @@
-import { type ROUTES } from '@/core/const';
-
-export type Route = typeof ROUTES;
-
-export type RouteValues = Exclude<Route[keyof Route], '*' | '/' | 'docs/en' | 'docs/ru'>;
